+++ src/base/b-router/b-router.ts
@@ -1164,13 +1164,14 @@ export default class bRouter extends iData {
 	protected async onLink(e: MouseEvent): Promise<void> {
 		const
 			a = <HTMLElement>e.delegateTarget,
-			href = a.getAttribute('href');
+			href = a.getAttribute('href')?.trim();
 
 		const cantPrevent =
 			!this.interceptLinks ||
 			href == null ||
 			href === '' ||
 			href.startsWith('#') ||
+			href.startsWith('javascript:') ||
 			isExternal.test(href);
 
 		if (cantPrevent) {