+++ webpack.config.js
@@ -37,8 +37,8 @@ async function buildFactory(entry, buildId) {
 		resolveLoader: await include('build/resolve-loader.webpack'),
 		externals: await include('build/externals.webpack')({buildId}),
 
-		plugins: [...plugins.values()],
 		module: {...modules, rules: [...modules.rules.values()]},
+		plugins: [...plugins.values()],
 
 		mode: webpack.mode(),
 		optimization: await include('build/optimization.webpack')({buildId, plugins}),
@@ -46,6 +46,7 @@ async function buildFactory(entry, buildId) {
 		devtool: await include('build/devtool.webpack'),
 		cache: await include('build/cache.webpack')({buildId}),
 		watchOptions: include('build/watch-options.webpack'),
+		snapshot: include('build/snapshot.webpack'),
 
 		...await include('build/other.webpack')({buildId})
 	};