+++ src/base/b-bottom-slide/CHANGELOG.md
@@ -9,6 +9,24 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.112 (2020-12-18)
+
+#### :boom: Breaking Change
+
+* Now `forceInnerRender` is toggled to `true` by default
+
+## v3.0.0-rc.112 (2020-12-18)
+
+#### :bug: Bug Fix
+
+* Fixed `updateWindowPosition` with lazy rendering
+
+## v3.0.0-rc.111 (2020-12-16)
+
+#### :bug: Bug Fix
+
+* Fixed render logic
+
 ## v3.0.0-rc.110 (2020-12-16)
 
 #### :rocket: New Feature