+++ src/base/b-router/CHANGELOG.md
@@ -23,6 +23,18 @@ Changelog
 
 * Improved documentation
 
+## v3.0.0-rc.118 (2020-12-24)
+
+#### :bug: Bug Fix
+
+* Trim href-s before go
+
+## v3.0.0-rc.113 (2020-12-18)
+
+#### :bug: Bug Fix
+
+* Fixed handling of `javascript:` links
+
 ## v3.0.0-rc.110 (2020-12-16)
 
 #### :rocket: New Feature