+++ src/core/router/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.114 (2020-12-22)
+
+#### :bug: Bug Fix
+
+* Fixed clearing of the route history
+
 ## v3.0.0-rc.110 (2020-12-16)
 
 #### :boom: Breaking Change